requirements.txt
beautifulsoup4==4.11.2
discord.py==2.1.1
keyboard==0.13.5
PyYAML==6.0
requests==2.27.1
